Delete unnecessary variable "retries"
The purpose of this variable was to track the number of retries to search for a partner with whom the random user has not worked.

We shuffle the elements instead of putting the cursor in a random position, in that sense we can be sure that all the possible combinations have been reviewed when the cursor is out of range, since the search is carried out in a linear way.